1) Changed percentage to int. 2) Changed the logic of matchDetails for vendor percentage calculation because the previous logic itself was wrong. 3) Added Unit Test for Processor layer of Vendor and Match Engine to make sure the logic is correct

+++ src/main/java/com/clm/matching/models/VendorMatchOverviewResponseDTO.java
@@ -14,6 +14,6 @@ public class VendorMatchOverviewResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private double matchPercentage;
+    private int matchPercentage;
     private List<OptionMatchResponseDTO> options;
 }
